test: update tests for auth [temp]
